+++ api-examples/pretrain-tlm.py
@@ -39,8 +39,6 @@ def train():
     parser.add_argument("--basedir", type=str)
     parser.add_argument("--train_file", type=str, required=True, help='File path to use for train file')
     parser.add_argument("--valid_file", type=str, required=True, help='File path to use for valid file')
-    parser.add_argument("--dataset_cache", type=str, default=os.path.expanduser('~/.bl-data'),
-                        help="Path or url of the dataset cache")
     parser.add_argument("--dataset_key", default="reddit",
                         help="dataset key for basedir")
     parser.add_argument("--embed_type", type=str, default='default',
@@ -87,17 +85,12 @@ def train():
                         type=int,
                         default=-1,
                         help="Local rank for distributed training (-1 means use the environment variables to find)")
-    parser.add_argument("--chars_per_word",
-                        type=int,
-                        default=40,
-                        help="How many max characters per word")
 
     args = parser.parse_args()
 
     if args.basedir is None:
         args.basedir = 'lm-{}-bpe-{}'.format(args.dataset_key, os.getpid())
     logging.basicConfig(level=logging.INFO if args.local_rank in [-1, 0] else logging.WARN)
-    logger.info("Cache directory [%s]", args.dataset_cache)
 
     num_gpus = int(os.environ.get("WORLD_SIZE", 1))
     args.distributed = args.distributed or num_gpus > 1
