+++ layers/eight_mile/pytorch/layers.py
@@ -3378,27 +3378,6 @@ def init_distributed(local_rank):
     torch.distributed.init_process_group(backend='nccl', init_method='env://')
     return device
 
-
-class Average:
-    def __init__(self, name, fmt=':f'):
-        self.name = name
-        self.fmt = fmt
-        self.val = 0
-        self.avg = 0
-        self.sum = 0
-        self.count = 0
-
-    def update(self, val, n=1):
-        self.val = val
-        self.sum += val * n
-        self.count += n
-        self.avg = self.sum / self.count
-
-    def __str__(self):
-        fmtstr = '{name} {val' + self.fmt + '} ({avg' + self.fmt + '})'
-        return fmtstr.format(**self.__dict__)
-
-
 class SingleHeadReduction(nn.Module):
     """
     Implementation of the "self_attention_head" layer from the conveRT paper (https://arxiv.org/pdf/1911.03688.pdf)
