+++ layers/eight_mile/utils.py
@@ -1622,3 +1622,24 @@ def print_table(rows: collections.namedtuple, columns: Optional[Set[str]] = None
         for field in fields:
             print("%*s = %s" % (hwidth, field, getattr(row, field)))
 
+
+@export
+class Average:
+    def __init__(self, name, fmt=':f'):
+        self.name = name
+        self.fmt = fmt
+        self.val = 0
+        self.avg = 0
+        self.sum = 0
+        self.count = 0
+
+    def update(self, val, n=1):
+        self.val = val
+        self.sum += val * n
+        self.count += n
+        self.avg = self.sum / self.count
+
+    def __str__(self):
+        fmtstr = '{name} {val' + self.fmt + '} ({avg' + self.fmt + '})'
+        return fmtstr.format(**self.__dict__)
+
